+++ src/main/java/com/wuit/dx/service/ProductService.java
@@ -12,5 +12,5 @@ public interface ProductService {
 
     List<Product> getAllProducts();
 
-    Product getProductById(Long id);
+    Product getProductById(int id);
 }
